refactor: separate welcome & level content, store in separate dirs
